Move addons to their own directory.
Addons that have built code or docker images include Grafana and
Servicegraph. These are not a part of Mixer. These are built and
included with the default release of Istio.
