+++ addons/grafana/mixer-dashboard.json
