+++ mixer/deploy/kube/README.md
@@ -1,3 +1,7 @@
+All the configuration below and in this directory is for test/dev of
+Mixer and potentially out of date. Official Istio install files are
+under [`/insall/kubernetes`](/install/kubernetes).
+
 # Istio Mixer K8s Configuration
 
 These configurations stand up an instance of Mixer, a Prometheus instance that scrapes metrics Mixer exposes, and
@@ -46,9 +50,10 @@ The Grafana UI is exposed on port `3000`. This configuration stores Grafana's da
 configure to be an ephemeral directory. To persist your Grafana configuration for longer than the lifetime of the `grafana`
 pod, map this directory to a real volume. Note that it is possible to export your Grafana config and re-import it later.
 
-An example grafana dashboard is provided in `conf/grafana-dashboard.json`. It can
-be imported into a running grafana instance that has a `prometheus` datasource
-configured at `http://prometheus:9090/`.
+An example grafana dashboard is provided in
+[`/addons/grafana/grafana-dashboard.json`](/addons/grafana/grafana-dashboard.json). It
+can be imported into a running grafana instance that has a
+`prometheus` datasource configured at `http://prometheus:9090/`.
 
 
 If you're using minikube to test these deployments locally, get to the UI by running:
